Fix fatal errors being invisible when running on Linux
The Linux back end has no implementation for OS_MsgBox yet, which causes
fatal errors to not be shown anywhere, not even in the console. This
fixes that by implementing it using SDL2, which is already used. This
should also handle properly displaying this on X11 or Wayland, instead
of having to deal with this ourselves.