Fix case-sensitiveness (fix #175, #178 and #179)
On case-sensitive OSes, a lot of files are not properly loaded because
of case mismatches. Instead of hardcoding the paths, a new header file
(Platform/Path.h) now contains all the relevant strings.

Issue #178 is fixed in the sense that now loading a previous game on
Linux results in a crash instead of doing nothing: the game was not
loaded because of a case-sensitive mismatch on the `Saves` folder.