Lower the language standard requirement to C++20
It was bumped to C++23 originally because of std::to_underlying, but
since then it was discovered that it's not available with the latest
Android toolchain.  Many popular GNU/Linux and *BSD systems are also
not ready for it, so reduce to C++20 while we're not using any C++23
features yet.