Clean indoor-related rendering
Regarding loop detection - it is bad enough if we ever get back to
the same sector as it means we would be adding the same faces again,
usually ad infinitum. But it seems this is not true in general.
The comment is inline to keep this matter in mind.

viewing_portal_id had no use.

Document some funky variables.

Reflowed some lines to ease with line breakpoints debugging.