Fix minimap construction
Fixes #1831.

We only allow to have the close-to-portal hack/fallback to happen
when frustum would be built nonetheless.
This ensures we do not render areas behind closed portals on the
minimap.
This also simplifies the code considerably, removing duplication
of new node construction, and ensures we always run the loop check.

The assert was removed as it was redundant with C++ array checks.

This commit also moves the minimap marking early to behave more
like vanilla. This also allows to restore the direction check
to its original place, saving some sky rendering and checks.

This works almost like in vanilla but is slightly more conservative,
e.g., it does not render final corridor walls on the minimap of
initial Castle Harmondale which are actually hidden behind the
inplaced walls which do not have a minimap line (in vanilla
the minimap suggests the corridor is longer than it seems,
and in OE it suggests it is unfinished - either is ultimately
wrong but much less critical than the fixed issue).

Other than that, OE now renders minimap fragments at the exact
same time vanilla does. Checked this in the most problematic areas:
Castle Harmondale, Temple of Light, Mercenary Guild, Celeste, The Pit.