Add preliminary support for MSVC Clang
This adds support for MSVC-compatible Clang, clang-cl.

Sadly, clang-cl supports a hybrid of cl's and clang's options [1]
but not all of either.
Our dependencies do not know how to pass correct flags.
This commit fixes passing of some of OE's own flags but there are
still many unknown (passed but should not) and missing (not passed
but should) ones.
The compilation output is messy but the resulting binary seems
to generally work; debugging in Visual Studio works too.
Tests do not work because they set `-Werror,-Wunknown-argument`
or `-Werror=unknown-argument`.

`OE_BUILD_COMPILER` is dropped per agreement in PR #1860.

Additionally, this commit moves the linker picking into the actual
gcc/clang path as it is not used elsewhere and is then effectively
confusing.

[1] https://clang.llvm.org/docs/UsersManual.html