Fix parsing of EVENT_ReceiveDamage
The `who` field needs to be set as it is used when interpreted.
Previously, it was left from some other event that set it, as the compiler (at least vc) reused the same memory for the new event.

This commit also adds a defensive zeroing of the event data (not great but better than nothing for now).

Additional event-relevant NOTEs and TODOs are added/updated as well.

Fixes #1898