Got rid of the copies of the viewport values in ViewingParams.
This required slightly changing some things in Engine - both MM7Initialization and FinalInitialization set the viewport to the viewingparams values, and depended on those being set in MM6_Initialize. I changed that to just do the same thing as OpenGLRenderer, but tbh both MM7Initialization and FinalInitialization probably shouldn't be doing this at all. Deleting both of the SetViewport calls there doesn't break anything in the game, but it might be required by tests?