Minor renderer cleanup (#2163)
* Unified initialization of the viewport

* Replaced usage of engine->config in BaseRenderer and OpenGLRenderer, as Renderer has it's own pointer to the same config.

This allowed removing a if (firstInit) check in BaseRenderer, as that was really a check for "if the engine has been assigned", which happens after renderer has been initialized.

* Moved the viewport update to BaseRenderer::updateRenderDimensions. This makes it be called from both NullRenderer::Initialize and NullRenderer::Reinitialize.

This is kinda clunky, but the OpenGLRenderer calls Reinitialize from Initialize, and the whole game was depending on that happening also in the NullRenderer case, which was failing tests.